Coverage in gitignore, multiple vcs
The .coverage has been added to the gitignore and also the application
halts with exception in case of no or multiple virtual private clouds.
This because if you have more than one virtual private cloud you may
create the security group in the wrong place, and now this is not
possible anymore. In the future, in case of multiple virtual private
clouds, user may have an option to choose which vpc he desires to
create the security group.